pyproject.toml
[project]
name = "news"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "feedparser>=6.0.12",
    "orjson>=3.11.3",
    "pandas>=2.3.2",
    "pyarrow>=21.0.0",
    "python-dateutil>=2.9.0.post0",
    "pyyaml>=6.0.3",
    "trafilatura>=2.0.0",
]
